[Enterprise Search] Refactor shared SchemaErrorsAccordion component (#99069)
* Move SchemaErrorsAccordion to its own folder

+ give it its own export

* Move errors-specific copy to its own file

+ Capitalize ID heading

* [UI polish] Accordion header tweaks

- Fix responsive behavior
- Make Review button more accessible (a within a button is not semantically correct)
- Add <code> markup for field type

* [UI polish] CSS tweaks

- use OOTB EUI CSS classes for borders (see https://elastic.github.io/eui/#/layout/accordion#styled-for-forms)
- tweak spacing
- misc fixes

* [Code polish] Misc cleanup

- Simplify / DRY out props types
- Prefer unique IDs/names over map indexes for React keys
- Improve var names

* Simplify conditional "view" column + misc fixes

- Rename prop to match AS's generateSomePath helpers
- Remove need for itemId - AS should be able to generate its own URL route without it
- Fix accessibility for view column - should be totally hidden to screen readers if not present
- Fix semantics of view link - should be an <a> link not a button

* Update WS's use of SchemaErrorsAccordion
