+++ x-pack/plugins/enterprise_search/public/applications/shared/schema/errors_accordion/constants.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+
+export const ERROR_TABLE_ID_HEADER = i18n.translate(
+  'xpack.enterpriseSearch.schema.errorsTable.heading.id',
+  { defaultMessage: 'ID' }
+);
+
+export const ERROR_TABLE_ERROR_HEADER = i18n.translate(
+  'xpack.enterpriseSearch.schema.errorsTable.heading.error',
+  { defaultMessage: 'Error' }
+);
+
+export const ERROR_TABLE_REVIEW_CONTROL = i18n.translate(
+  'xpack.enterpriseSearch.schema.errorsTable.control.review',
+  { defaultMessage: 'Review' }
+);
+
+export const ERROR_TABLE_VIEW_LINK = i18n.translate(
+  'xpack.enterpriseSearch.schema.errorsTable.link.view',
+  { defaultMessage: 'View' }
+);
