+++ x-pack/plugins/enterprise_search/public/applications/shared/schema/errors_accordion/schema_errors_accordion.scss
@@ -1,0 +1,26 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+.schemaErrorsAccordion {
+  // Fixes the accordion button content not stretching to full-width
+  .euiAccordion__button > :last-child {
+    flex-grow: 1;
+  }
+
+  // Tweak spacing
+  .euiAccordion__childWrapper {
+    position: relative;
+    top: -($euiSizeS);
+    padding-left: $euiSizeL;
+    padding-right: $euiSizeL;
+  }
+
+  // Remove extra border-bottom on last table row
+  .euiTableRow:last-child .euiTableRowCell {
+    border-bottom: none;
+  }
+}
