+++ x-pack/plugins/enterprise_search/public/applications/shared/schema/index.ts
@@ -8,3 +8,4 @@
 export { SchemaAddFieldModal } from './add_field_modal';
 export { SchemaFieldTypeSelect } from './field_type_select';
 export { SchemaErrorsCallout } from './errors_callout';
+export { SchemaErrorsAccordion } from './errors_accordion';
