--- x-pack/plugins/enterprise_search/public/applications/shared/schema/schema_errors_accordion.scss
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-.schemaFieldError {
-  border-top: 1px solid $euiColorLightShade;
-
-  &:last-child {
-    border-bottom: 1px solid $euiColorLightShade;
-  }
-
-  // Something about the EuiFlexGroup being inside a button collapses the row of items.
-  // This wrapper div was injected by EUI and had 'with: auto' on it.
-  .euiIEFlexWrapFix {
-    width: 100%;
-  }
-}
